% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggiornaStudentiVotiModuli.R
\name{aggiornaStudentiVotiModuli}
\alias{aggiornaStudentiVotiModuli}
\title{Una funzione che aggrega output di aggiornaStudentiVoti.}
\usage{
aggiornaStudentiVotiModuli(basiBiologiche, biologiaMolecolare, geneticaUmana)
}
\arguments{
\item{basiBiologiche, }{output di aggiornaStudentiVoti per Basi Biologiche, update studenti.}

\item{biologiaMolecolare, }{output di aggiornaStudentiVoti per Biologia Molecolare, update studenti.}

\item{geneticaUmana, }{output di aggiornaStudentiVoti per Genetica Umana, update studenti.}
}
\value{
Il file tab delimited derivato da studenti_voti.txt,  riempito con i risultati di uno dei tre moduli.
}
\description{
Una funzione che aggrega output aggiornaStudentiVoti per i tre moduli. La matrice dei voti e' vuota ma sono inseriti tutti gli studenti indipendentemente dall'anno di corso.
}
\examples{
\dontrun{
aggiornaStudentiVotiModuli(basiBiologiche=bb.set, 
                          biologiaMolecolare=bm.set, 
                          geneticaUmana=gu.set)

}

}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, University of Torino, Italy
}
