% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggiungiRisultatiModulo.R
\name{aggiungiRisultatiModulo}
\alias{aggiungiRisultatiModulo}
\title{Una funzione che aggrega output di aggiornaStudentiVoti.}
\usage{
aggiungiRisultatiModulo(excel.esame, input.voti, output.voti,
  modulo = c("BB", "BM", "GU"))
}
\arguments{
\item{excel.esame, }{e' l'excel scaricato per un modulo dalle prove parziali su SS3.}

\item{input.voti, }{output di aggiornaStudentiVotiModuli.}

\item{output.voti, }{output di aggiornaStudentiVoti per Genetica Umana, update studenti.}

\item{modulo, }{selezione del modulo per il quale i voto vanno aggiunti}
}
\value{
Il file tab delimited vuoto, studenti_voti.txt,  ma con l'inserimento di tutti gli studenti indipendentemente dall'anno di corso.
}
\description{
Una funzione che aggrega output aggiornaStudentiVoti per i tre moduli.
}
\examples{
\dontrun{

aggiungiRisultatiModulo(excel.esame="basi_biologiche.xls", 
                        input.voti="studenti_voti.txt",
                        output.voti="studenti_votiBB.txt", modulo="BB")
aggiungiRisultatiModulo(excel.esame="biologia_molecolare.xls", 
                        input.voti="studenti_votiBB.txt",
                        output.voti="studenti_votiBBBM.txt", modulo="BM")
aggiungiRisultatiModulo(excel.esame="genetica_umana.xls", 
                        input.voti="studenti_votiBBBM.txt",
                        output.voti="studenti_votiBBBMGU.txt", modulo="GU")

}

}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, University of Torino, Italy
}
