% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbalizza.R
\name{verbalizza}
\alias{verbalizza}
\title{Una funzione che selezione gli studenti da verbalizzare.}
\usage{
verbalizza(excel.verbalizzazione, voti.aggregati, voti.verbalizzabili,
  studenti.missing)
}
\arguments{
\item{excel.verbalizzazione, }{file excel della prenotazione per appello verbalizzante.}

\item{voti.aggregati, }{file tab delimited con i voti aggregati generati da aggregaVotiModul.}

\item{voti.verbalizzabili, }{tab delimited output con solo i voti aggregati per gli studenti registratiall'appello verbalizzante.}

\item{studenti.missing, }{tab delimited output con la matrice del file di verbalizzazione per gli studenti che non trovo nei miei dati.}
}
\value{
Il file tab delimited  riempito con i risultati parziali ed il voto integrato per gli studenti da verbalizzare.
}
\description{
Una funzione che seleziona dal output di aggregaVotiModuli il set di studenti registrati per la verbalizzazione.
}
\examples{
\dontrun{
verbalizza(excel.verbalizzazione="verbalizzante.xls", 
                  voti.aggregati="canaleA_studenti_votiBBBMGU_aggregati_190227.txt",
                  voti.verbalizzabili="canaleA_studenti_da_verbalizzare.txt", 
                  studenti.missing="canaleA_studenti_da_verbalizzare_missing.txt")
}

}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, University of Torino, Italy
}
