% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggiornaStudentiVoti.R
\name{aggiornaStudentiVoti}
\alias{aggiornaStudentiVoti}
\title{Una funzione che crea una lista voti.}
\usage{
aggiornaStudentiVoti(excel.esame, elenco.studenti)
}
\arguments{
\item{excel.esame, }{e' l'excel scaricato per un modulo dalle prove parziali su SS3}

\item{elenco.studenti, elenco.studenti}{e' l'elenco degli studenti dell'anno in corso per un certo canale.}
}
\value{
Un data.frame aggiornato con i nuovi studenti. Importante questo data.frame e' completamente vuoto
}
\description{
Una funzione che crea una lista voti vuota e la aggiorna per la presenza di studenti non presenti nella lista dell'anno in corso.
}
\examples{
\dontrun{
bb.set <- aggiornaStudentiVoti(excel.esame="basi_biologiche.xls", 
                                elenco.studenti="../canaleA_studenti.txt")
bm.set <- aggiornaStudentiVoti(excel.esame="biologia_molecolare.xls", 
                                elenco.studenti="../canaleA_studenti.txt")
gu.set <- aggiornaStudentiVoti(excel.esame="genetica_umana.xls", 
                                elenco.studenti="../canaleA_studenti.txt")

}

}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, University of Torino, Italy
}
